app/src/test/java/com/example/cupcake/OrderViewModelTests.kt
package com.example.cupcake

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import com.example.cupcake.model.OrderViewModel
import org.junit.Assert.assertEquals
import org.junit.Rule
import org.junit.Test

class OrderViewModelTests {

    @get:Rule
    var instantTaskExecutorRule = InstantTaskExecutorRule()

    @Test
    fun quantity_twelve_cupcakes() {
        val viewModel = OrderViewModel()
        viewModel.quantity.observeForever {}
        viewModel.setQuantity(12)
        assertEquals(12, viewModel.quantity.value)
    }

    @Test
    fun price_twelve_cupcakes() {
        val viewModel = OrderViewModel()
        viewModel.price.observeForever {}
        viewModel.setQuantity(12)
        assertEquals("$24.00", viewModel.price.value)
    }

    @Test
    fun price_same_day_delivery() {
        val viewModel = OrderViewModel()
        viewModel.price.observeForever {}
        viewModel.setQuantity(12)
        viewModel.setPickupDate(viewModel.pickupDateOptions[0])
        assertEquals("$27.00", viewModel.price.value)
    }

    @Test
    fun price_not_same_day_delivery() {
        val viewModel = OrderViewModel()
        viewModel.price.observeForever {}
        viewModel.setQuantity(12)
        viewModel.setPickupDate(viewModel.pickupDateOptions[1])
        assertEquals("$24.00", viewModel.price.value)
    }
}